refactor() remove legacy code, fix conversion to be memory-aware
