--- src/core/services/loading/loading_old/file_loader/augmented_ingestors.py
@@ -1,228 +1,0 @@
-"""
-memory_optimized_ingestors.py
-Memory-efficient batch generators with integrated monitoring and cleanup.
-"""
-import pyarrow.parquet as pq
-import csv
-import gc
-from typing import Iterable, List, Tuple, Optional, Any, Union
-from .....setup.logging import logger
-
-
-def batch_generator_parquet(
-    path: str, 
-    headers: List[str], 
-    chunk_size: int = 20_000,
-    memory_monitor: Optional[Any] = None
-) -> Iterable[List[Tuple]]:
-    """
-    Memory-efficient Parquet batch generator with integrated monitoring.
-    
-    Key optimizations:
-    - Lazy column mapping
-    - Minimal string conversion
-    - Aggressive cleanup between batches
-    - Memory pressure monitoring
-    """
-    pf = None
-    name_to_idx = None
-    batch_count = 0
-    
-    try:
-        pf = pq.ParquetFile(path)
-        
-        for record_batch in pf.iter_batches(batch_size=chunk_size):
-            # Memory check before processing
-            if memory_monitor and memory_monitor.should_prevent_processing():
-                logger.error(f"Memory limit exceeded during Parquet processing at batch {batch_count}")
-                raise MemoryError("Memory limit exceeded during Parquet processing")
-            
-            # Build column mapping only once
-            if name_to_idx is None:
-                name_to_idx = {
-                    record_batch.schema.field(i).name: i 
-                    for i in range(record_batch.num_columns)
-                }
-                logger.debug(f"Parquet column mapping established: {len(name_to_idx)} columns")
-            
-            # Use vectorized approach - convert columns once per batch
-            try:
-                # Convert to dict once (faster than per-cell .as_py())
-                batch_dict = record_batch.to_pydict()
-                
-                # Build column arrays in header order
-                cols = []
-                for h in headers:
-                    if h in name_to_idx and h in batch_dict:
-                        cols.append(batch_dict[h])
-                    else:
-                        cols.append([None] * record_batch.num_rows)
-                
-                # Transpose columns into rows (much faster than nested loops)
-                batch_rows = []
-                for row_vals in zip(*cols):
-                    # Preserve types - only stringify if absolutely needed
-                    batch_rows.append(tuple(row_vals))
-                
-                # Yield before cleanup
-                yield batch_rows
-                batch_count += 1
-                
-            finally:
-                # Aggressive cleanup after each batch
-                del batch_rows
-                del record_batch
-                
-                # Periodic aggressive cleanup
-                if batch_count % 3 == 0:
-                    gc.collect()
-                    if memory_monitor:
-                        if memory_monitor.is_memory_pressure_high():
-                            cleanup_stats = memory_monitor.perform_aggressive_cleanup()
-                            logger.info(f"Batch {batch_count}: Memory cleanup freed {cleanup_stats.get('freed_mb', 0):.1f}MB")
-                
-    finally:
-        # Final cleanup
-        if pf:
-            del pf
-        if name_to_idx:
-            del name_to_idx
-        gc.collect()
-
-
-def batch_generator_csv(
-    path: str, 
-    headers: List[str], 
-    chunk_size: int = 20_000,
-    encoding: str = 'utf-8',
-    memory_monitor: Optional[Any] = None
-) -> Iterable[List[Tuple]]:
-    """
-    Memory-efficient CSV batch generator with integrated monitoring.
-    
-    Key optimizations:
-    - Smaller dialect detection sample
-    - Streaming processing
-    - Memory pressure checks
-    - Efficient row normalization
-    """
-    batch_count = 0
-    total_rows = 0
-    
-    try:
-        with open(path, 'r', newline='', encoding=encoding, errors='replace') as f:
-            # Use smaller sample for dialect detection to save memory
-            sample_size = min(2048, chunk_size * 25)  # Reduced from 4096
-            sample = f.read(sample_size)
-            f.seek(0)
-            
-            # Detect dialect
-            try:
-                dialect = csv.Sniffer().sniff(sample)
-            except Exception:
-                dialect = csv.excel()
-                dialect.delimiter = ';'
-            
-            del sample  # Free sample memory immediately
-            
-            reader = csv.reader(f, dialect=dialect)
-            first_row = next(reader, None)
-            
-            # Header detection with minimal memory impact
-            skip_header = False
-            if first_row:
-                if any(header in first_row for header in headers):
-                    skip_header = True
-                elif all(not cell.strip().isdigit() for cell in first_row if cell.strip()):
-                    skip_header = True
-            
-            if not skip_header and first_row:
-                f.seek(0)
-                reader = csv.reader(f, dialect=dialect)
-            
-            # Process with memory monitoring
-            batch = []
-            headers_len = len(headers)
-            
-            for row in reader:
-                # Periodic memory checks (every 1000 rows)
-                if total_rows % 1000 == 0 and memory_monitor:
-                    if memory_monitor.should_prevent_processing():
-                        logger.error(f"Memory limit exceeded at row {total_rows}")
-                        raise MemoryError(f"Memory limit exceeded at row {total_rows}")
-                
-                # Efficient row normalization
-                row_len = len(row)
-                if row_len < headers_len:
-                    # Extend row efficiently
-                    row.extend([None] * (headers_len - row_len))
-                elif row_len > headers_len:
-                    # Truncate row efficiently
-                    row = row[:headers_len]
-                
-                batch.append(tuple(row))
-                total_rows += 1
-                
-                # Yield batch when size reached
-                if len(batch) >= chunk_size:
-                    yield batch
-                    batch_count += 1
-                    batch = []  # Clear reference
-                    
-                    # Memory cleanup every few batches
-                    if batch_count % 5 == 0:
-                        gc.collect()
-                        if memory_monitor and memory_monitor.is_memory_pressure_high():
-                            cleanup_stats = memory_monitor.perform_aggressive_cleanup()
-                            logger.info(f"CSV batch {batch_count}: Cleanup freed {cleanup_stats.get('freed_mb', 0):.1f}MB")
-            
-            # Yield remaining rows
-            if batch:
-                yield batch
-    
-    finally:
-        # Final cleanup
-        gc.collect()
-
-
-def create_batch_generator(
-    path: str,
-    headers: List[str],
-    chunk_size: int = 20_000,
-    encoding: str = 'utf-8',
-    memory_monitor: Optional[Any] = None,
-    file_format: Optional[str] = None
-) -> Iterable[List[Tuple]]:
-    """
-    Factory function for memory-aware batch generators.
-    
-    Args:
-        path: File path
-        headers: Expected column headers
-        chunk_size: Rows per batch
-        encoding: File encoding for CSV
-        memory_monitor: Memory monitor instance
-        file_format: Force format ('csv' or 'parquet'), or None for auto-detect
-        
-    Returns:
-        Memory-efficient batch generator
-    """
-    if file_format is None:
-        # Auto-detect format
-        if path.lower().endswith('.parquet'):
-            file_format = 'parquet'
-        else:
-            file_format = 'csv'
-    
-    logger.info(f"Creating memory-aware {file_format.upper()} generator for {path}")
-    
-    if file_format == 'parquet':
-        return batch_generator_parquet(
-            path, headers, chunk_size, memory_monitor
-        )
-    elif file_format == 'csv':
-        return batch_generator_csv(
-            path, headers, chunk_size, encoding, memory_monitor
-        )
-    else:
-        raise ValueError(f"Unsupported file format: {file_format}")
